Add old implementation files for line-breaking. (BROKEN)
My original attempt to solve the line-breaking issues had certain pluses,
so I copied those class here in an attempt to build the new implementation
with the hard-fought knowledge from the first attempt.

Not all files in this check-in compile. Do not checkpoint here unless you
know what you're doing.