Mavenize Flying Saucer
Initial check-in of POM file to Maveniz Flying Saucer.  This attempts to
create a package that it identical to the one offered by the Ant build
without changing the structure of the project in any way.

*The SWT jar must be installed to your local repo.* See POM.xml for
details and sample command line.

The right thing to do for complete Mavenization would be to restructure
the project to fall inline with Maven build principals.